pyproject.toml
[project]
name = "bonsai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "langgraph>=0.2",
    "langchain-openai>=0.3",
    "tavily-python>=0.5",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "datasets>=3.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["slow: marks tests as slow (real LLM + Tavily calls)"]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
